scxtop: adding profilin_events folder
