+++ tools/scxtop/src/lib.rs
@@ -14,12 +14,11 @@ pub mod cpu_stats;
 pub mod edm;
 mod event_data;
 mod keymap;
-mod kprobe_event;
 mod llc_data;
 pub mod mangoapp;
 mod node_data;
-mod perf_event;
 mod perfetto_trace;
+pub mod profiling_events;
 pub mod protos;
 mod search;
 mod stats;
@@ -36,13 +35,12 @@ pub use cpu_stats::CpuStatTracker;
 pub use event_data::EventData;
 pub use keymap::Key;
 pub use keymap::KeyMap;
-pub use kprobe_event::available_kprobe_events;
-pub use kprobe_event::KprobeEvent;
 pub use llc_data::LlcData;
 pub use node_data::NodeData;
-pub use perf_event::available_perf_events;
-pub use perf_event::PerfEvent;
 pub use perfetto_trace::PerfettoTraceManager;
+pub use profiling_events::{
+    available_kprobe_events, available_perf_events, KprobeEvent, PerfEvent, ProfilingEvent,
+};
 pub use protos::*;
 pub use search::Search;
 pub use stats::StatAggregation;
@@ -65,7 +63,7 @@ pub const TRACE_FILE_PREFIX: &str = "scxtop_trace";
 pub const STATS_SOCKET_PATH: &str = "/var/run/scx/root/stats";
 pub const LICENSE: &str = "Copyright (c) Meta Platforms, Inc. and affiliates.
 
-This software may be used and distributed according to the terms of the 
+This software may be used and distributed according to the terms of the
 GNU General Public License version 2.";
 pub const SCHED_NAME_PATH: &str = "/sys/kernel/sched_ext/root/ops";
 
@@ -130,28 +128,6 @@ impl FilteredEventState {
         self.count = 0;
         self.scroll = 0;
         self.selected = 0;
-    }
-}
-
-#[derive(Clone, Debug)]
-pub enum ProfilingEvent {
-    Perf(PerfEvent),
-    Kprobe(KprobeEvent),
-}
-
-impl ProfilingEvent {
-    pub fn event_name(&self) -> &str {
-        match self {
-            ProfilingEvent::Perf(p) => p.event_name(),
-            ProfilingEvent::Kprobe(k) => &k.event_name,
-        }
-    }
-
-    pub fn value(&mut self, reset: bool) -> anyhow::Result<u64> {
-        match self {
-            ProfilingEvent::Perf(p) => p.value(reset),
-            ProfilingEvent::Kprobe(k) => k.value(reset),
-        }
     }
 }
