+++ tools/scxtop/src/profiling_events/perf.rs
